OpenViduCall 2.14.0

+++ frontend/openvidu-webcomponent-build.js
@@ -3,71 +3,63 @@ const concat = require('concat');
 const VERSION = require('./package.json').version;
 
 // Fixed app-route bug: https://github.com/angular/angular/issues/24674
-module.exports.prepareWebcomponent = function () { 
+module.exports.prepareWebcomponent = function () {
   console.log("Preparing webcomponent files ...");
   const appModule = './src/app/app.module.ts';
   replaceText(appModule, "bootstrap: [AppComponent]", "// bootstrap: [AppComponent]");
 }
 
-module.exports.buildWebcomponent = function () {
+module.exports.buildWebcomponent = async () => {
   console.log("Building OpenVidu Web Component (" + VERSION + ")");
+  const tutorialWcPath = '../../openvidu-tutorials/openvidu-webcomponent/web';
+  const e2eWcPath = '../webcomponent-test-e2e/web';
 
-  buildElement()
-  .then(() => {
-    copyFiles().then(() => {
-      return restore();
-    });
-  })
-  .then(() => {
+  try {
+    await buildElement();
+    await copyFiles(tutorialWcPath);
+    await copyFiles(e2eWcPath);
+    await restore();
     console.log('OpenVidu Web Component (' + VERSION + ') built');
-  }).catch((error) => console.error(error));
+  } catch (error) {
+    replaceText(appModule, "// bootstrap: [AppComponent]", "bootstrap: [AppComponent]");
+    console.error(error);
+  }
 }
 
 async function buildElement() {
   const files = [
-    './dist/openvidu-teaching/runtime.js',
-    './dist/openvidu-teaching/polyfills.js',
-    './dist/openvidu-teaching/scripts.js',
-    './dist/openvidu-teaching/main.js',
+    './dist/openvidu-call/runtime.js',
+    './dist/openvidu-call/polyfills.js',
+    './dist/openvidu-call/scripts.js',
+    './dist/openvidu-call/main.js',
   ];
 
   try {
     await fs.ensureDir('openvidu-webcomponent');
     await concat(files, './openvidu-webcomponent/openvidu-webcomponent-' + VERSION + '.js')
-    await fs.copy('./dist/openvidu-teaching/styles.css', './openvidu-webcomponent/openvidu-webcomponent-' + VERSION + '.css');
+    await fs.copy('./dist/openvidu-call/styles.css', './openvidu-webcomponent/openvidu-webcomponent-' + VERSION + '.css');
   } catch (err) {
-    console.error('Error executing build funtion in webcomponent-builds.js', err);
-    replaceText(appModule, "// bootstrap: [AppComponent]", "bootstrap: [AppComponent]");
+    console.error('Error executing build function in webcomponent-builds.js');
+    throw err;
   }
 }
 
-async function copyFiles() {
-  const destination = '../../../openvidu-tutorials/openvidu-webcomponent/web';
-  try {
-    console.log("Copying openvidu-webcomponents files from: ./openvidu-webcomponent to: " + destination);
-    await fs.ensureDir('openvidu-webcomponent');
-    await fs.copy('./openvidu-webcomponent/', destination);
-  } catch (err) {
-    console.error('Error executing copy function in webcomponent-builds.js', err);
-    replaceText(appModule, "// bootstrap: [AppComponent]", "bootstrap: [AppComponent]");
-  }
-}
-
-async function deleteFilesBuilt() {
-  try {
-    console.log("Deleting openvidu-webcomponent directory...");
-    await fs.ensureDir('openvidu-webcomponent');
-    await fs.remove("./openvidu-webcomponent");
-  } catch (err) {
-    console.error('Error executing delete function in webcomponent-builds.js', err);
-    replaceText(appModule, "// bootstrap: [AppComponent]", "bootstrap: [AppComponent]");
+async function copyFiles(destination) {
+  if (fs.existsSync(destination)) {
+    try {
+      console.log("Copying openvidu-webcomponent files from: ./openvidu-webcomponent to: " + destination);
+      await fs.ensureDir('openvidu-webcomponent');
+      await fs.copy('./openvidu-webcomponent/', destination);
+    } catch (err) {
+      console.error('Error executing copy function in webcomponent-builds.js');
+      throw err;
+    }
   }
 }
 
 async function restore() {
   const appModule = './src/app/app.module.ts';
   replaceText(appModule, "// bootstrap: [AppComponent]", "bootstrap: [AppComponent]");
-  deleteFilesBuilt();
 }
 
 function replaceText(file, originalText, changedText) {
@@ -83,4 +75,3 @@ function replaceText(file, originalText, changedText) {
   });
 
 }
-
